requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
dlib>=19.24.0
numpy>=1.21.0
requests>=2.26.0
pillow>=9.0.0
opencv-python-headless>=4.5.0
